add notif for killswitch activation
